ecommerce-wishlist/requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.30.0
click==8.3.0
colorama==0.4.6
fastapi==0.104.1
git-filter-repo==2.47.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic_core==2.14.1
python-dotenv==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.14
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.24.0
httpx==0.27.2
pydantic-settings  # ✅ ƏLAVƏ ET


